fix :  secure delete controller

+++ controllers/sauce.js
@@ -108,12 +108,24 @@ exports.likeDislike = (req, res, next) => {
 exports.deleteSauce = (req, res, next) => {
   Sauce.findOne({ _id: req.params.id })
     .then((sauce) => {
+      if (!sauce) {
+        res.status(404).json({
+          error: new Error("Sauce inexistante"),
+        });
+      }
+      if (sauce.userId !== req.auth.userId) {
+        res.status(401).json({
+          error: new Error("Requête non authorisée"),
+        });
+      } else {
+        fs.unlink(`images/${filename}`, () => {
+          Sauce.deleteOne({ _id: req.params.id })
+            .then(() => res.status(200).json({ message: "Sauce supprimée !" }))
+            .catch((error) => res.status(400).json({ error }));
+        });
+      }
+
       const filename = sauce.imageUrl.split("/images/")[1];
-      fs.unlink(`images/${filename}`, () => {
-        Sauce.deleteOne({ _id: req.params.id })
-          .then(() => res.status(200).json({ message: "Sauce supprimée !" }))
-          .catch((error) => res.status(400).json({ error }));
-      });
     })
     .catch((error) => res.status(500).json({ error }));
 };
